refactor: reorganizar estrutura de pastas para melhor modularização.
